Structs are in headers
